+++ src/main/java/com/jaehyun/reservation/user/review/domain/dto/ReviewReqDto.java
@@ -1,0 +1,17 @@
+package com.jaehyun.reservation.user.review.domain.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ReviewReqDto {
+  private double stars;
+  private String reviewText;
+}
